Refactor date handling in filters and metrics to ensure month is returned as an integer

+++ app/Nova/Filters/BulanFilter.php
@@ -53,6 +53,6 @@ class BulanFilter extends Filter
 
     public function default()
     {
-        return $this->default == 'cm' ? date('m') : $this->default;
+        return $this->default == 'cm' ? (int) date('m') : $this->default;
     }
 }
